// relays_indexer/src/settings/types.ts
export type Settings = {
    initial_pubkey: string;
    pubkey_index: number;
    pubkeys_per_process: number;
    max_fetch_events: number;
    relay_index: number;
    relays_connections: number;
    relays_betch_size: number;
    indexer_interval: number
    pubkeys_per_notes: number;
    max_fetch_notes: number;
}